Move ogg into durations
