+++ proteus-lib/src/container/info/duration/ogg.rs
